feat!: Add Function and Optimizer traits for general optimization
Although the primary goal of Gomez - solving non-linear systems of
equations - remains still the same, I believe that providing means for
general numerical optimization may benefit someone for the features that
Gomez has: first-class support for bounds, simple, low-level, iterative
API and various utilities. Moreover, two out of three algorithms
implemented so far are optimization algorithms, not solvers anyway.

As solving equations is the primary goal, the information about general
optimization is rather omitted from the documentation and that is on
purpose.