feat!: Require contiguous storage for x values
In my code, I often need to pass x to an evalution procedure that takes
&[f64] slice. Therefore I had to collect the values to a newly allocated
Vec in order to achieve that. To me, requiring contiguous storage for x
is reasonable limitation as it usually is a vector.

For fx, contiguous storage is not required at the moment as I haven't
encountered a situation where I needed it.