feat: Update nalgebra to v0.31
This would be normally considered a breaking change. But I am going to
yank version 0.3.0 of Gomez due to messed up publish (I accidentaly
published the crate with a binary). So the 0.3.1 version will be
essentially first version in 0.3.x branch, and that is why this commit
is not marked as breaking change.