feat!: Remove Function and Optimizer from the prelude
Having both Solver and Optimizer traits in the prelude, when a type
implements both, the compiler can't distinguish between them for a call
to `next` method. Given that function optimization is not the main goal
of gomez, but just a by-product, as a temporary solution
function-related traits are removed from the prelude.

However, the design of the traits should be somehow improved so that the
clash is not there at all.